requirements.txt
pyaudio
pynput
numpy
faster-whisper
rumps
pyobjc
py2app
python-dotenv
setuptools
huggingface_hub